python/requirements.txt
sqlalchemy
mysqlclient
marshmallow-sqlalchemy
mysql-connector-python
fastapi
uvicorn
orjson 
pandas 